requirements.txt
fastapi
uvicorn
google-genai
httpx
sentence-transformers
pydantic
bcrypt
PyJWT
python-multipart
python-dotenv
supabase
langchain-core
langchain-community
langgraph
langchain-ollama
neo4j
